Sort lists and change default monster inititative
The lists for battle themes and adding a monster to the encounter were not sorted, making it difficult to find what you are looking for. Monsters also defaulted to an initiative of 0, which is invalid for the options in the combobox. The lists are now sorted and the default initiatives have been set to 1